Move 42Auth button to separate component
